Rules mixin to remove some clutter from Quoridor class with minor quoridor file cleanup
