+++ src/constants.py
@@ -17,3 +17,5 @@ WHITE = (255, 255, 255)
 BLACK = (0, 0, 0)
 GRAY = (200, 200, 200)
 SEMI_BLACK = (0, 0, 0, 128)
+DEFAULT_FONT_SIZE = 32
+WALL_EDGE_COORD = x[-1]
